+++ notification.js
@@ -10,24 +10,7 @@ window.addEventListener("DOMContentLoaded", () => {
         <div class="card-body">
           ${data.body}
         </div>
-        <div class="card-footer">
-          ${data.buttons ? data.buttons.map((button) => `<button type="button" class="btn btn-${button.class}" id="btn-${button.id}">${button.html}</button>`).join("") : ``}
-        </div>
       </div>`);
-
-    // Add click events for buttons
-    if (data.buttons) {
-      window.requestAnimationFrame(() => {
-        data.buttons.map((button) => {
-          $(`#btn-${button.id}`).click((e) => {
-            if (typeof button.click === 'function') {
-              button.click(e);
-            }
-            window.close();
-          });
-        });
-      })
-    }
 
     // Flash notification if specified
     if (data.flash) {